feat+refactor: add basic SplitAny, restructure the project closer to release form
Non-basic implementation will take quite a bit of effort due to lack of *both* span.IndexOfAny(ROS<segment>) (for various reasons) and SearchValues<byte[]>.

In addition, implementing SplitAny<T> where T is char, rune or another U8String will also be quite complicated because for chars and runes it will need to do converstion in place, and, yet again, limit the split iterator to being a ref struct, forcing manual implementation of the methods to make it usable beyond simple foreach.

I think it's best to just stop and leave the features in the current state and focus on finishing other parts of implementation + test coverage. Further work on splitting will require a refactor/rewrite to source generation and/or templating to work around C#'s type system limitations and hard requirements for performance+UX.